+++ src/store/index.ts
@@ -1,9 +1,21 @@
 import Vue from "vue";
 import Vuex from "vuex";
 import { version, authors, dependencies } from "../../package.json";
+import players from "@/assets/json/players.json";
 import scytheFactions from "@/assets/json/scythe-factions.json";
 import scythePlays from "@/assets/json/scythe-plays.json";
-import { Game, ScytheFaction, ScythePlay } from "@/types";
+import tapestryCivilizations from "@/assets/json/tapestry-civilizations.json";
+import tapestryPlays from "@/assets/json/tapestry-plays.json";
+import tapestryScenarios from "@/assets/json/tapestry-scenarios.json";
+import {
+  Game,
+  Player,
+  ScytheFaction,
+  ScythePlay,
+  TapestryCivilization,
+  TapestryPlay,
+  TapestryScenario,
+} from "@/types";
 
 Vue.use(Vuex);
 
@@ -19,8 +31,36 @@ export default new Vuex.Store({
 
     games: [] as Game[],
 
+    players: players as Player[],
+
     scythePlays: scythePlays as ScythePlay[],
     scytheFactions: scytheFactions as ScytheFaction[],
+
+    tapestryPlays: tapestryPlays.map((play) => {
+      play.players.map((playerFromPlay) => {
+        if (playerFromPlay.civilizationSlug) {
+          playerFromPlay.civilization = tapestryCivilizations.find(
+            (civ) => civ.slug === playerFromPlay.civilizationSlug
+          );
+          delete playerFromPlay.civilizationSlug;
+        }
+        return Object.assign(
+          playerFromPlay,
+          players.find(
+            (playerFromList) => playerFromPlay.id === playerFromList.id
+          ) || {}
+        );
+      });
+      if (play.scenarioId) {
+        play.scenario = tapestryScenarios.find(
+          (scenario) => play.scenarioId === scenario.id
+        );
+        delete play.scenarioId;
+      }
+      return play;
+    }) as TapestryPlay[],
+    tapestryCivilizations: tapestryCivilizations as TapestryCivilization[],
+    tapestryScenarios: tapestryScenarios as TapestryScenario[],
   },
 
   mutations: {